feat: word selection
