+++ src/features/game-play/index.ts
@@ -1,3 +1,1 @@
 export { GamePlay } from "./ui/GamePlay";
-export type { Position } from "./model/types";
-export { useDraggableSelection } from "./lib/useDraggableSelection";
