+++ src/features/word-selection/index.ts
@@ -1,0 +1,2 @@
+export { useDraggableSelection } from "./lib/useDraggableSelection";
+export { SelectableLettersGrid } from "./ui/SelectableLettersGrid";
